Pull resetDatabase into its own file
...and also make it callable in server-side tests.
